+++ tests/std_stream.rs
@@ -9,11 +9,11 @@ fn stream_type_is_generic_over_item_type() {
         import std.cli { Args }
         import std.stream { Stream }
 
-        node passthrough(stream: Stream[Int]) -> out: Stream[Int] {
+        node passthrough(stream: Stream[i64]) -> out: Stream[i64] {
             $stream -> $out
         }
 
-        program main(args: Args) -> exit_code: Int {
+        program main(args: Args) -> exit_code: i64 {
             0 -> $exit_code
         }
     "#;
@@ -28,7 +28,7 @@ fn std_stream_copy_file_runs_without_bytes_materialization() {
         import std.fs { open_file, copy_to_file }
         import std.seq { head, tail }
 
-        program main(args: Args) -> exit_code: Faultable[Int] {
+        program main(args: Args) -> exit_code: Faultable[i64] {
             $args -> argv -> $paths
             $paths -> head -> $input_path
             $paths -> tail -> head -> $output_path
@@ -67,7 +67,7 @@ fn std_stream_open_file_can_stream_contents_to_seq() {
         import std.seq { head }
         import std.stream as stream
 
-        program main(args: Args) -> exit_code: Faultable[Int] {
+        program main(args: Args) -> exit_code: Faultable[i64] {
             $args -> argv -> head -> $input_path
             $input_path -> open_file -> stream.to_seq -> concat_bytes -> $contents
             [$contents, "\n"] -> concat_bytes -> $output
@@ -101,7 +101,7 @@ fn std_stream_read_at_reads_a_slice_without_consuming_the_stream() {
         import std.io { write_stdout }
         import std.seq { head }
 
-        program main(args: Args) -> exit_code: Faultable[Int] {
+        program main(args: Args) -> exit_code: Faultable[i64] {
             $args -> argv -> head -> $input_path
             $input_path -> open_file -> $stream
             ($stream, 6, 6) -> read_at -> $middle
